+++ src/navigation/Main/MainNavigator.tsx
@@ -1,7 +1,7 @@
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import React from 'react';
 import { MAIN_ROUTES } from '~navigation/Main/mainTypes';
-import { HomeScreen } from '~screens';
+import { EntityScreen, HomeScreen } from '~screens';
 import { HomeHeader } from '~components';
 
 const Stack = createNativeStackNavigator();
@@ -15,6 +15,7 @@ export const MainNavigator = () => {
           component={HomeScreen}
           options={{ headerShown: true, header: (props) => <HomeHeader {...props} /> }}
         />
+        <Stack.Screen name={MAIN_ROUTES.ENTITY} component={EntityScreen} />
       </Stack.Group>
     </Stack.Navigator>
   );